+++ R/buildmxPedigrees.R
@@ -420,3 +420,21 @@ fitPedigreeModel <- function(
   }
   fitted_model
 }
+
+
+#' Align Phenotype Vector to Matrix Format for OpenMx
+#'
+#' This function takes a pedigree data frame, a specified phenotype column, and a vector of IDs to keep, and returns a matrix formatted for use in OpenMx models. The resulting matrix has one row and columns corresponding to the specified IDs, with values taken from the phenotype column of the pedigree.
+#' @param ped A data frame representing the pedigree, containing at least the columns specified by \code{phenotype} and \code{personID}.
+#' @param phenotype A character string specifying the column name in \code{ped} that
+#'  contains the phenotype values to be aligned.
+#' @param keep_ids A vector of IDs for which the phenotype values should be extracted and aligned. These IDs should correspond to the values in the \code{personID} of \code{ped}.
+#' @param personID A character string specifying the column name in \code{ped} that contains the individual IDs. Default is "ID".
+#' @export
+
+
+alignPhenToMatrix <- function(ped, phenotype, keep_ids, personID = "ID") {
+  obs_ids <- make.names(as.character(keep_ids))
+  pheno_vals <- ped[[phenotype]][match(as.character(keep_ids), as.character(ped[[personID]]))]
+  matrix(as.double(pheno_vals), nrow = 1, dimnames = list(NULL, obs_ids))
+}